requirements.txt
Pillow>=9.0.0
tkinterdnd2-universal>=0.4.3